Small changes in code and CLI

+++ src/cli.rs
@@ -16,13 +16,13 @@ pub use clap::{Parser, Subcommand};
     long_about = None,
 )]
 pub struct Cli {
-    /// Interval between commands, ignored if --run_times is 1. For format see https://docs.rs/humantime/2.1.0/humantime/fn.parse_duration.html.
+    /// Interval between commands, ignored if --run_times/-n is 1. For format see https://docs.rs/humantime/2.1.0/humantime/fn.parse_duration.html.
     #[arg(short = 'I', long)]
     pub interval: Option<humantime::Duration>,
     /// How many times to run the command. Specifying 0 will cause commands to run infinitely until
     /// the user manually terminates the program.
     #[arg(short = 'n', long, default_value_t = 1, verbatim_doc_comment)]
-    pub run_times: u32,
+    pub run_times: u64,
     /// Delimiter used for separating responses. Also used by `list-cpus` and `list-sensors` commands.
     #[arg(short, long, default_value = "\n")]
     pub delimiter: String,
